requirements.txt
openai
langchain==0.1.20
streamlit
langchain-openai
python-dotenv
pypdf
faiss-cpu
langchainhub
PyPDF2==3.0.1
pymongo==4.7.3
